MakeFile: adjustment, AddressUtil: Fix base58 encoding with leading zeros, CudaKeySearchDevice: rename round function to prevent error with CUDA library

+++ CLKeySearchDevice/CLKeySearchDevice.cpp
@@ -209,7 +209,7 @@ void CLKeySearchDevice::init(const secp256k1::uint256 &start,
 		} else {
 			p = secp256k1::multiplyPoint(
 					secp256k1::uint256(
-							(uint64_t) _threads * _blocks * _pointsPerThread)
+							(uint64_t) (_threads * _blocks * _pointsPerThread))
 							* _stride, g);
 		}
 
@@ -221,9 +221,9 @@ void CLKeySearchDevice::init(const secp256k1::uint256 &start,
 
 void CLKeySearchDevice::doStep() {
 	try {
-		uint64_t numKeys = (uint64_t) _blocks * _threads * _pointsPerThread;
+		uint64_t numKeys = (uint64_t) (_blocks * _threads * _pointsPerThread);
 
-		if (!_randomMode && _iterations < 2 && _start.cmp(numKeys) <= 0) {
+		if (!_randomMode && _iterations < (uint64_t) 2 && _start.cmp(numKeys) <= (uint64_t) 0) {
 			_stepKernelWithDouble->call(_blocks, _threads, _pointsPerThread,
 					_compression, _chain, _x, _y, _xInc, _yInc,
 					_deviceTargetList.ptr, _deviceTargetList.size,
